# src/back/Imolt.Api/contracts/openapi.yaml
...
openapi: 3.0.3
info:
  title: ИМОЛТ — расчёт вывоза и утилизации отходов
  version: 1.2.0
  description: |
    Договор расчётной части сервиса ИМОЛТ: расчёт стоимости вывоза и
    утилизации строительных отходов по Москве и Московской области, подбор
    полигонов, маршрут, коммерческое предложение и заявка.

    **Как читать.** Операции сгруппированы по трём предметным областям
    расчётной части — «Справочники», «Расчёт», «Сделка». Границы областей
    объявлены записью ADR-0001, и группировка здесь — их внешняя проекция:
    по метке `x-область` видно, какая часть системы отвечает за операцию.

    **Что уже работает.** Отвечают только `/health` и `/ready`. Остальные
    операции помечены `x-состояние: объявлено`: форма запроса и ответа
    зафиксирована, сервер их пока не обслуживает. Мок поднимается из этого же
    файла, примеры ответов взяты из раздела 7 запроса на дизайн и согласованы
    между собой (адрес «Москва, ул. Годовикова, 9», дата актуальности
    17.09.2026).

    **Чего в договоре нет.** Поведение, не определённое заказчиком, не
    додумано: такие места помечены `x-вопросы` со ссылкой на открытый вопрос.
    Отложенные требования (R-047, R-062 — R-066) в версию 1 не входят.

    **Соглашения.**

    - Пути и имена полей — латиницей, тексты и описания — по-русски;
      идентификаторы кода остаются латиницей по правилам проекта.
    - Денежная сумма передаётся строкой с двумя знаками после точки: двоичная
...
      в макете (R-059).
  contact:
...
  license:
...
servers:
...
security:
  - {}
  - bearerAuth: []
paths:
...
  /v1/waste-groups:
    get:
      operationId: listWasteGroups
      tags: [Справочники]
      summary: Справочник групп отходов
      description: |
        Список групп отходов с поиском по названию и коду каталога отходов. Отдаёт цену
        перевозки за тонна-километр и коэффициент плотности: по ним
        интерфейс показывает пересчёт меры, а расчёт — стоимость.
      x-область: справочники
      x-требования: [R-013, R-039, R-017, R-015]
      x-экран: [Э-01, Э-02, Э-11]
      x-состояние: реализовано
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: страница справочника
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WasteGroupPage' }
              examples:
                twoWasteGroups:
                  summary: поиск по строке «лом»
                  value:
                    items:
                      - id: beton-lom
                        name: Лом бетона и железобетона
                        fkkoCodes: ['8 22 201 01 21 5']
                        transportPricePerTonKm: { amount: '12.00', currency: RUB }
                        densityTonPerCubicMeter: 2.0
                        updatedAt: '2026-09-17'
                      - id: kirpich-lom
                        name: Лом кирпичной кладки
                        fkkoCodes: ['8 23 101 01 21 5']
                        transportPricePerTonKm: { amount: '14.00', currency: RUB }
                        densityTonPerCubicMeter: 1.5
                        updatedAt: '2026-09-17'
                    total: 2
                    limit: 10
                    offset: 0
        '400': { $ref: '#/components/responses/BadRequest' }
  /v1/waste-groups/{wasteGroupId}:
    parameters:
      - $ref: '#/components/parameters/WasteGroupIdPath'
    get:
      operationId: getWasteGroup
      tags: [Справочники]
      summary: Карточка группы отходов
      x-область: справочники
      x-требования: [R-039]
      x-экран: [Э-11]
      x-состояние: реализовано
      responses:
        '200':
          description: карточка группы
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WasteGroup' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateWasteGroup
      tags: [Справочники]
      summary: Изменить цену перевозки и плотность группы
      description: |
        Редактор цен менеджера данных. Цена перевозки вводится за
        тонна-километр по группе и умножается на километраж в расчёте —
        второго места, где она живёт, нет.

        Кто вправе править, определяет политика доступа; она пока не
        установлена (Q-011), поэтому договор объявляет требование сессии,
        а не роль.
      security:
        - bearerAuth: []
      x-область: справочники
      x-требования: [R-042, R-043, R-048]
      x-вопросы: [Q-011, Q-013]
      x-экран: [Э-11]
      x-состояние: объявлено
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WasteGroupUpdate' }
            example:
              transportPricePerTonKm: { amount: '32.00', currency: RUB }
      responses:
        '200':
          description: группа изменена, дата актуальности обновлена
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WasteGroup' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableContent' }
  /v1/landfills:
    get:
      operationId: listLandfills
      tags: [Справочники]
      summary: Реестр полигонов
      description: |
        Полигоны с принимаемыми группами отходов, тарифами утилизации,
        статусом и датой его актуальности. Режим «Справочник цен» в интерфейсе
        и редактор данных читают один и тот же список.
      x-область: справочники
      x-требования: [R-040, R-004, R-028, R-048]
      x-экран: [Э-03, Э-11]
      x-состояние: реализовано
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - name: wasteGroupId
          in: query
          description: оставить полигоны, принимающие эту группу отходов
          schema: { type: string }
        - name: status
          in: query
          description: оставить полигоны с этим статусом
          schema: { $ref: '#/components/schemas/LandfillStatus' }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: страница реестра
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LandfillPage' }
              examples:
                singleLandfill:
                  summary: фильтр по группе «Лом бетона»
                  value:
                    items:
                      - id: vostok-timohovo
                        name: Комплекс переработки «Восток»
                        legalEntity: ООО «Восток»
                        address: Московская обл., Богородский г. о., д. Тимохово
                        coordinates: { latitude: 55.7286, longitude: 38.2153 }
                        status: active
                        statusUpdatedAt: '2026-09-17'
                        tariffs:
                          - wasteGroupId: beton-lom
                            disposalPricePerTon: { amount: '450.00', currency: RUB }
                            updatedAt: '2026-09-17'
                    total: 1
                    limit: 10
                    offset: 0
        '400': { $ref: '#/components/responses/BadRequest' }
  /v1/landfills/{landfillId}:
    parameters:
      - $ref: '#/components/parameters/LandfillIdPath'
    get:
      operationId: getLandfill
      tags: [Справочники]
      summary: Карточка полигона
      description: |
        Карточка хранит историю смены юридического лица: тарифы и статусы
        при смене владельца не теряются (R-041).
      x-область: справочники
      x-требования: [R-040, R-041]
      x-экран: [Э-11]
      x-состояние: реализовано
      responses:
        '200':
          description: карточка полигона
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LandfillCard' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/landfills/{landfillId}/reviews:
    parameters:
      - $ref: '#/components/parameters/LandfillIdPath'
    get:
      operationId: listLandfillReviews
      tags: [Справочники]
      summary: Отзывы и оценки полигона
      description: |
        Оценки пользователей — обратная связь по актуальности данных, а не
        рейтинг услуги: расхождение цены или статуса с действительностью
        видно раньше, чем его заметит менеджер данных (риск AR-002).

        Экрана под эту возможность в запросе на дизайн пока нет.
      x-область: справочники
      x-требования: [R-031]
      x-состояние: реализовано
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: страница отзывов
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LandfillReviewPage' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createLandfillReview
      tags: [Справочники]
      summary: Оставить оценку и отзыв о полигоне
      security:
        - bearerAuth: []
      x-область: справочники
      x-требования: [R-031, R-048]
      x-вопросы: [Q-011]
      x-состояние: реализовано
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LandfillReviewInput' }
            example:
              rating: 2
              text: тариф на месте оказался 520 ₽/т вместо 480 ₽/т
      responses:
        '201':
          description: отзыв принят
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LandfillReview' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableContent' }
  /v1/address-suggestions:
    get:
      operationId: suggestAddresses
      tags: [Справочники]
      summary: Подсказки адреса вывоза
      description: |
        Подсказки приходят через сервер, а не из браузера: ключ доступа к
        внешней службе остаётся на сервере (R-056, риск AR-006). Принимаются
        только адреса Москвы и Московской области и только с координатами —
        без них расстояние не посчитать.
      x-область: справочники
      x-требования: [R-012, R-055, R-056]
      x-вопросы: [Q-014]
      x-экран: [Э-01]
      x-состояние: реализовано
      parameters:
        - name: query
          in: query
          required: true
          description: начало адреса, не короче трёх символов
          schema: { type: string, minLength: 3, maxLength: 200 }
          example: Годовикова 9
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: подсказки, пригодные для расчёта
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AddressSuggestionPage' }
              example:
                items:
                  - id: dadata-77-godovikova-9
                    value: г Москва, ул Годовикова, д 9
                    coordinates: { latitude: 55.8055, longitude: 37.6206 }
                    area: moscow
                total: 1
                limit: 10
                offset: 0
        '400': { $ref: '#/components/responses/BadRequest' }
        '503': { $ref: '#/components/responses/UpstreamUnavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Маркер сессии из операции `createSession`. Расчёт и коммерческое
        предложение задуманы доступными без входа; окончательный перечень
        закрытых операций — открытый вопрос Q-011.
  parameters:
    SearchQuery:
      name: query
      in: query
      description: 'поиск по названию; для групп отходов — ещё и по коду каталога отходов'
      required: false
      schema: { type: string, maxLength: 200 }
    Limit:
      name: limit
      in: query
      description: |
        Сколько записей вернуть. По умолчанию 10: первый экран результатов
        показывает десять полигонов, остальные догружаются (R-029, R-060).
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
    Offset:
      name: offset
      in: query
      description: сколько записей пропустить
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    WasteGroupIdPath:
      name: wasteGroupId
      in: path
      required: true
      description: идентификатор группы отходов
      schema: { type: string }
      example: beton-lom
    LandfillIdPath:
      name: landfillId
      in: path
      required: true
      description: идентификатор полигона
      schema: { type: string }
      example: vostok-timohovo
...
  responses:
    BadRequest:
      description: запрос не разобран или параметр вне допустимых значений
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:imolt:problem:validation'
            title: Запрос не прошёл проверку
            status: 400
            detail: Параметр limit должен быть от 1 до 100
    NotFound:
      description: запись не найдена
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:imolt:problem:not-found'
            title: Запись не найдена
            status: 404
    Unauthorized:
      description: операция требует входа
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:imolt:problem:authentication-required'
            title: Нужен вход
            status: 401
    Forbidden:
      description: |
        Сессия есть, но операция закрыта подпиской или ролью. Перечень
        закрытых операций не установлен — Q-011.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:imolt:problem:subscription-required'
            title: Детали доступны по подписке
            status: 403
    UnprocessableContent:
      description: запрос разобран, но нарушает правило предметной области
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:imolt:problem:allocation-mismatch'
            title: Распределение не сходится с объёмом
            status: 422
            detail: По группе «Лом бетона» распределено 18 т из 20 т
            errors:
              - field: entries[1].quantity.value
                message: увеличьте на 2 т либо уменьшите объём группы
...
    UpstreamUnavailable:
      description: |
        Внешняя служба расстояний или подсказок не ответила либо исчерпан
        лимит обращений (риск AR-007). Сохранённые данные при этом остаются
        доступны: отказ источника не равен отказу обслуживания.
      headers:
        Retry-After:
          description: через сколько секунд повторить
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:imolt:problem:distance-service-unavailable'
            title: Не удалось рассчитать расстояния
            status: 503
            detail: Повторите попытку
...
  schemas:
    Money:
      type: object
      description: |
        Денежная сумма. Передаётся строкой с двумя знаками после точки:
        двоичная дробь округляет рубли по дороге, а смета этого не прощает.
        Форматирование по локали ru-RU — работа интерфейса (R-061).
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: '^-?[0-9]+\.[0-9]{2}$'
          example: '19800.00'
        currency:
          type: string
          enum: [RUB]
      additionalProperties: false
...
    Coordinates:
      type: object
      description: координаты точки; без них расстояние по дорожной сети не считается
      required: [latitude, longitude]
      properties:
        latitude: { type: number, format: double, minimum: -90, maximum: 90 }
        longitude: { type: number, format: double, minimum: -180, maximum: 180 }
      additionalProperties: false
    ServiceArea:
      type: string
      description: зона обслуживания, к которой относится адрес
      enum: [moscow, moscowRegion]
...
    Problem:
      type: object
      description: |
        Описание ошибки по RFC 9457. Поле `type` — устойчивый код причины:
        интерфейс ветвится по нему, а не по тексту заголовка.
      required: [type, title, status]
      properties:
        type: { type: string, example: 'urn:imolt:problem:validation' }
        title: { type: string, example: Запрос не прошёл проверку }
        status: { type: integer, example: 400 }
        detail: { type: string }
        instance: { type: string, description: 'путь запроса, на котором возникла ошибка' }
        errors:
          type: array
          description: разбор по полям — для форм
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
    Page:
      type: object
      description: общая часть страничного ответа
      required: [total, limit, offset]
      properties:
        total: { type: integer, description: сколько записей всего подходит под запрос }
        limit: { type: integer }
        offset: { type: integer }
    WasteGroup:
      type: object
      description: |
        Группа отходов справочника: название, коды ФККО, цена перевозки за
        тонна-километр и коэффициент плотности для пересчёта меры.
      required: [id, name, fkkoCodes, transportPricePerTonKm, densityTonPerCubicMeter, updatedAt]
      properties:
        id: { type: string, example: beton-lom }
        name: { type: string, example: Лом бетона и железобетона }
        fkkoCodes:
          type: array
          items: { type: string }
          example: ['8 22 201 01 21 5']
        transportPricePerTonKm: { $ref: '#/components/schemas/Money' }
        densityTonPerCubicMeter: { type: number, format: double, example: 2.0 }
        updatedAt: { type: string, format: date, example: '2026-09-17' }
    WasteGroupPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/WasteGroup' }
    WasteGroupUpdate:
      type: object
      description: изменяется то, что передано; непереданное остаётся прежним
      minProperties: 1
      properties:
        name: { type: string }
        fkkoCodes:
          type: array
          items: { type: string }
        transportPricePerTonKm: { $ref: '#/components/schemas/Money' }
        densityTonPerCubicMeter: { type: number, format: double, minimum: 0 }
      additionalProperties: false
    LandfillStatus:
      type: string
      description: |
        Состояние полигона: принимает отходы, заблокирован либо не подтверждён.
        `unconfirmed` — когда сообщение источника не распознано и статус не
        подтверждён вручную.
      enum: [active, blocked, unconfirmed]
    LandfillTariff:
      type: object
      required: [wasteGroupId, disposalPricePerTon, updatedAt]
      properties:
        wasteGroupId: { type: string, example: beton-lom }
        disposalPricePerTon: { $ref: '#/components/schemas/Money' }
        updatedAt: { type: string, format: date, example: '2026-09-17' }
...
    Landfill:
      type: object
      required: [id, name, address, coordinates, status, statusUpdatedAt, tariffs]
      properties:
        id: { type: string, example: vostok-timohovo }
        name: { type: string, example: Комплекс переработки «Восток» }
        legalEntity: { type: string, example: ООО «Восток» }
        address: { type: string }
        coordinates: { $ref: '#/components/schemas/Coordinates' }
        status: { $ref: '#/components/schemas/LandfillStatus' }
        statusUpdatedAt: { type: string, format: date }
        tariffs:
          type: array
          description: тарифы утилизации по принимаемым группам отходов
          items: { $ref: '#/components/schemas/LandfillTariff' }
    LandfillPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Landfill' }
    LandfillCard:
      allOf:
        - $ref: '#/components/schemas/Landfill'
        - type: object
          properties:
            legalEntityHistory:
              type: array
              description: смены юридического лица без потери тарифов и статусов (R-041)
              items:
                type: object
                required: [legalEntity, since]
                properties:
                  legalEntity: { type: string }
                  since: { type: string, format: date }
                  until: { type: string, format: date, nullable: true }
...
    AddressSuggestion:
      type: object
      required: [id, value, coordinates, area]
      properties:
        id: { type: string, description: идентификатор подсказки внешней службы }
        value: { type: string, example: 'г Москва, ул Годовикова, д 9' }
        coordinates: { $ref: '#/components/schemas/Coordinates' }
        area: { $ref: '#/components/schemas/ServiceArea' }
    AddressSuggestionPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/AddressSuggestion' }
...
    LandfillReviewInput:
      type: object
      required: [rating]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        text: { type: string, maxLength: 2000 }
      additionalProperties: false
    LandfillReview:
      type: object
      required: [id, landfillId, rating, createdAt]
      properties:
        id: { type: string }
        landfillId: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        text: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    LandfillReviewPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required: [items, averageRating]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/LandfillReview' }
            averageRating: { type: number, format: double, nullable: true }